requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
python-pptx==0.6.23
PyPDF2==3.0.1
pdf2image==1.16.3
pillow==10.2.0

# Google Gemini API
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3

# Development dependencies
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Optional: For advanced PDF processing
pdfplumber==0.10.3
pypdf==4.0.1

# For cost estimation
tiktoken==0.5.2